requirements.txt
asgiref==3.8.1
bcrypt==4.2.1
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Jinja2==3.1.5
MarkupSafe==3.0.2
mysqlclient==2.2.6
prettytable==3.12.0
pyecharts==2.0.7
PyJWT==1.7.1
simplejson==3.19.3
sqlparse==0.5.3
typing_extensions==4.12.2
wcwidth==0.2.13
